fix compile warnings
git-svn-id: http://qon.jwdt.org/var/svn/catracker@385 97746f1a-7bc6-4ed5-8832-28bd89886fd8

+++ GPU/TPCCAGPUTracking/tracking-ca/standalone/cagpubuild/AliHLTTPCCAGPUTrackerOpenCL.cxx
@@ -336,7 +336,7 @@ int AliHLTTPCCAGPUTrackerOpenCL::InitGPU_Runtime(int sliceCount, int forceDevice
 
 	ocl->selector_events = new cl_event[fSliceCount];
 
-	HLTInfo("OPENCL Initialisation successfull (%d: %s %s (Frequency %d, Shaders %d) Thread %d, Max slices: %d, %d bytes used)", bestDevice, device_vendor, device_name, (int) freq, (int) shaders, fThreadId, fSliceCount, fGPUMemSize);
+	HLTInfo("OPENCL Initialisation successfull (%d: %s %s (Frequency %d, Shaders %d) Thread %d, Max slices: %d, %lld bytes used)", bestDevice, device_vendor, device_name, (int) freq, (int) shaders, fThreadId, fSliceCount, (long long int) fGPUMemSize);
 
 	return(0);
 }
